waylandsink: Destroy wl_callback when finish pipeline.
Wl_callback for the last frame is not freed if frame_redraw_callback()
is not called for the frame.

If wl_display_dispatch() is used for solving this problem, there is another
problem of freezing in interrupt shutdown such as ctrc + c.
This problem happens because waiting of dispatching lasts.

So directly destorying wl_callback is appropriate.